+++ src/instructions/moveq.cpp
@@ -34,7 +34,7 @@ std::string Moveq::disassembly(CPUState& cpu_state){
 
     std::ostringstream output;
     output << "moveq"
-           << " #$" << std::hex << this->imm_data
+           << " #$" << std::hex << (uint32_t)this->imm_data
            << ", " << DISASSEMBLER::effectiveAddress(this->dest_mode, this->dest_reg, this->data_size, cpu_state);
                
     return output.str();
